fix: refactor player manager
