+++ Assets/Scenes/GroupDetails.unity
@@ -1062,7 +1062,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: -63, y: -189}
+  m_AnchoredPosition: {x: -62.999985, y: -189}
   m_SizeDelta: {x: -152.91003, y: 23.769989}
   m_Pivot: {x: 0.5, y: 1}
 --- !u!114 &452374627
@@ -1200,7 +1200,6 @@ MonoBehaviour:
   groupMembersText: {fileID: 0}
   chatNameInput: {fileID: 102930394}
   startChatButton: {fileID: 1431679970}
-  disbandGroupButton: {fileID: 1191300837}
   editGroupButton: {fileID: 1134375408}
   createRoom: {fileID: 115949935}
   memberItemPrefab: {fileID: 3148322257433250143, guid: 4fabe633012922e4c89581fdf9ec25fe, type: 3}
@@ -2600,7 +2599,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &1191300836
 RectTransform:
   m_ObjectHideFlags: 0
@@ -3362,7 +3361,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_text: Delete Group
+  m_text: Leave Group
   m_isRightToLeft: 0
   m_fontAsset: {fileID: 11400000, guid: 8f586378b4e144a9851e7b34d9b748ee, type: 2}
   m_sharedMaterial: {fileID: 2180264, guid: 8f586378b4e144a9851e7b34d9b748ee, type: 2}
